Convert Messages page component to TS
